Moved logout handler into CharacterHandler
